+++ ampd/src/handlers/config.rs
@@ -57,11 +57,11 @@ pub enum Config {
     },
     StellarMsgVerifier {
         cosmwasm_contract: TMAddress,
-        http_url: Url,
+        rpc_url: Url,
     },
     StellarVerifierSetVerifier {
         cosmwasm_contract: TMAddress,
-        http_url: Url,
+        rpc_url: Url,
     },
     StacksMsgVerifier {
         cosmwasm_contract: TMAddress,
@@ -296,11 +296,11 @@ mod tests {
         let configs = vec![
             Config::StellarMsgVerifier {
                 cosmwasm_contract: TMAddress::random(PREFIX),
-                http_url: "http://localhost:8080/".parse().unwrap(),
+                rpc_url: "http://localhost:7545/".parse().unwrap(),
             },
             Config::StellarMsgVerifier {
                 cosmwasm_contract: TMAddress::random(PREFIX),
-                http_url: "http://localhost:8080/".parse().unwrap(),
+                rpc_url: "http://localhost:7545/".parse().unwrap(),
             },
         ];
 
@@ -313,11 +313,11 @@ mod tests {
         let configs = vec![
             Config::StellarVerifierSetVerifier {
                 cosmwasm_contract: TMAddress::random(PREFIX),
-                http_url: "http://localhost:8080/".parse().unwrap(),
+                rpc_url: "http://localhost:7545/".parse().unwrap(),
             },
             Config::StellarVerifierSetVerifier {
                 cosmwasm_contract: TMAddress::random(PREFIX),
-                http_url: "http://localhost:8080/".parse().unwrap(),
+                rpc_url: "http://localhost:7545/".parse().unwrap(),
             },
         ];