+++ ampd/src/handlers/stellar_verify_msg.rs
@@ -23,7 +23,7 @@ use voting_verifier::msg::ExecuteMsg;
 use crate::event_processor::EventHandler;
 use crate::handlers::errors::Error;
 use crate::handlers::errors::Error::DeserializeEvent;
-use crate::stellar::http_client::Client;
+use crate::stellar::rpc_client::Client;
 use crate::stellar::verifier::verify_message;
 use crate::types::TMAddress;
 
@@ -185,7 +185,7 @@ mod tests {
     use super::PollStartedEvent;
     use crate::event_processor::EventHandler;
     use crate::handlers::tests::{into_structured_event, participants};
-    use crate::stellar::http_client::Client;
+    use crate::stellar::rpc_client::Client;
     use crate::types::{EVMAddress, Hash, TMAddress};
     use crate::PREFIX;